% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{get_extr_request}
\alias{get_extr_request}
\title{Executes a custom CSOB AM GET request}
\usage{
get_extr_request(page = NULL, ...)
}
\arguments{
\item{...}{other arguments for the \pkg{httr}'s \code{\link[httr]{GET}} function}
}
\description{
And stores the HTML content
}
\examples{
get_extr_request()

}
