% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{get_last_page}
\alias{get_last_page}
\title{Returns number of pages that can be queried to get all offered funds}
\usage{
get_last_page(...)
}
\value{
an interger count
}
\description{
By default, around 20 funds per page are displayed on \url{https://www.csobam.cz/portal/asset-management/produkty-a-sluzby/aktualni-hodnoty-a-vykonnost}
}
\seealso{
base_site_URL for the URL that is being scrapped.
}
