% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{get_target_url_list}
\alias{get_target_url_list}
\title{Build URL list of queries}
\usage{
get_target_url_list(pageCount)
}
\arguments{
\item{pageCount}{- number of pages to scrap}
}
\description{
Build URL list of queries
}
