% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{get_targeted_URL_list}
\alias{get_targeted_URL_list}
\title{Build URL list of queries, in a table}
\usage{
get_targeted_URL_list(pageCount = NULL)
}
\arguments{
\item{pageCount}{- number of pages to scrap}
}
\description{
Build URL list of queries, in a table
}
