% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{returnURL}
\alias{returnURL}
\title{Return a website string for further use.}
\usage{

  returnURL(URL = "https://www.csobam.cz/portal/asset-management/produkty-a-sluzby/aktualni-hodnoty-a-vykonnost",
  ...)
}
\arguments{
\item{URL}{- The url that is used, e.g. where a desired table is located.
By default it is set to \url{https://www.csobam.cz/portal/asset-management/produkty-a-sluzby/aktualni-hodnoty-a-vykonnost}}
}
\value{
a URL string
}
\description{
If contains parameters, then they are appended.
}
