% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{get_extr_request}
\alias{get_extr_request}
\title{Prepare custom CSOB AM GET request}
\usage{
get_extr_request(url_req = returnURL())
}
\description{
And make it ready for HTML extraction purposes
}
\examples{
get_extr_request()

}
