% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{get_last_page}
\alias{get_last_page}
\title{How many pages do we have to query to get all funds managed by CSOB Asset Management.}
\usage{
get_last_page()
}
\value{
an interger count
}
\description{
Either hardcoded number (~ 24) or dynamic (here).
}
\seealso{
returnURL for the URL that is being scrapped.
}
