% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funds.R
\name{managed_funds}
\alias{managed_funds}
\title{Scrap all funds from a public CSOB AM website.}
\usage{
managed_funds(pages_to_scrap = "ALL", funds_to_scrap = NULL)
}
\arguments{
\item{pages_to_scrap}{- how many pages of available funds to scrap (by default ALL)}

\item{funds_to_scrap}{- how many funds (exact number) to scrap (by default NULL)}
}
\value{
a table that contains either all funds managed by CSOB AM (default setting) or their selected number
}
\description{
No processing is done at this stage (other than adding a proper header names)
}
\seealso{
returnURL for the URL that is being scrapped.
}
